Financial trait nerfed: +1 commerce only on tiles with a natural yield of at least 2 commerce or a total yield of at least 3 commerce. [advc.908a]
Bug fixed in v0.90 code: The AI no longer refuses to talk if the player brings in a war ally but isn't him-/herself at war with the AI. [advc.104i]

Refuse-to-talk duration after embargo reduced to 22 turns on average. (v0.89 had already reduced it from 30 to 25 turns.) [advc.130f]

Enabled the Unit AI routine for guarding high-yield tiles (in the code since v0.8) even when Barbarians are not set to Raging, but with some restrictions. [advc.300]

Minor improvements to the UWAI code for predicting nuclear war.

Slightly adjusted the size of the Spy unit models and of a few ships. [advc.002j]