A few misc. changes: BUG default settings changed (only affects new installations of AdvCiv): * Wonders tab, with color coding, enabled. [advc.004] * Sevopedia, sorted alphabetically, enabled. [advc.004] * Glance tab shows relations values by default, not Smilies. [advc.004] * Reduced line height of scoreboard. [advc.106d]
When naming a Great Person, the expected number of names skipped at random is now based on the number of civs in the game. [advc.005b]

All civs receive a research bonus if tech trading is disabled. [advc.550d]

Stopped-trading memory is deleted when a civ capitulates and when a capitulated vassal breaks free. [advc.130f]

Installation instructions in the manual revised.