Fixed problems with UN and AP resolutions not being recorded in the replay log. [advc.150b]
UWAI: Made the AI unwilling to talk for a while (same willing-to-talk logic as in BtS) after a force-war vote. Previously, UWAI often made peace a couple of turns after the vote. Also corrected an oversight in the treatment of defensive pacts: when a pact is triggered, the AI becomes unwilling to talk to _both_ war allies, and, upon making peace with one civ, the AI becomes willing to talk with any DP allies of that civ. [advc.104i]