Revised AI_foundValue (the function that determines where the AI settles) Still hardly tested [change id: advc.031 if none stated]
Fixed a bug I introduced in v0.90; the bug had caused K-Mod code to be ignored that discourages settling on top of a resource. While looking for the bug, I've improved said K-Mod code a bit (hopefully). Also fixed a related issue introduced around that time: when a resource could not yet be improved, no bonus yields had been counted for that resource. Now bonus yields are counted if the tech requirement for trading the resource is currently researchable.

Fixed a bug in K-Mod; code that was supposed to reduce found values in the proximity of other civs had had no effect. [advc.001]

Per-tile utility values are now sorted in descending order and summed up with decreasing weights, meaning that more weight is given to tiles that are likely to be worked early in the city's life. (Previously, I had achieved a similar effect by taking per-tile values to the power of 1.25, but this would also encourage cities with uniformly good tiles, e.g. 20 river plains, when it might be better to split that area between several cities.)

Some utility is counted for tiles in the radii of other cities, i.e. tiles that could be stolen from rival cities or shared with a friendly city.

Rearranged code to make the function easier to read and to avoid some unnecessary or counterproductive computations. (It's still messy though, both conceptually and in terms of implementation.)

Reconciled my own code for discouraging low-food sites and sites with too many bonus resources (advc.052) with BtS code serving a similar purpose.

Reduced the AI aversion for sites with numerous bad tiles (e.g. Tundra) a bit further.

In scenarios, the choice of the best city sites (function AI_updateCitySites) and the site that Settlers move to are now slightly randomized so that the AI doesn't necessarily pick the same spots for its first couple of cities in the same order every game. [advc.052]

Unrelated to found value:
The AI is now less likely to build Barracks before its first Settler. [advc.017]
Animals can't appear on Desert and Snow tiles unless there is freshwater. [advc.300]