A few usability tweaks (mostly)
When a goody hut would grant between 80% and 100% progress toward a tech, the whole tech is granted instead. [advc.314]

Decreased AI willingness to cancel trades with humans in large games because it's annoying to constantly renegotiate deals. [advc.036]

"Will import" alert (resource trades) only triggers if the importer is willing to give at least 3 gold per turn (previously 1 gold per turn). [advc.210e]

When the AI proposes a trade and offers more gold than it normally would, it no longer offers the same amount of gold when the human player negotiates the proposal. (Because the extra gold is supposed to be a reward for accepting the AI proposal.) [advc.026]

Increased the cost of Great Wall from 250 to 300. [advc.310]