Changed the way DeclareWarTradeRand (DWTR) is used by the AI.
This value from LeaderHead XML determines the probability of checking for war trades. The probability is now 10 / (DWTR + 10 * min{10,t}), where t is the number of turns that the war has lasted. For small t, the resulting probability is lower than in BtS, and, for large t, it's higher. I might set an individual DWTR value for each leader at a later point; currently they all have DWTR=40 as in BtS. [advc.161]

Revised code for change advc.104i (UWAI willing to talk only according to BtS rules when fighting a war for a third party). Now also applies when an AI civ is paid to start a war, which is what the manual had been saying all along.