AI improvements, mostly related to naval war (This commit contains some temporary code (marked with advc.tmp) that suppresses change advc.122, which is still work in progress.)
Minor (tentative) changes to the AI routines for picking up stranded units, and the game tries (but not very hard) to bump units to a landmass where their owner has a city. [advc.046]
Prompted by this CFC thread:
https://forums.civfanatics.com/threads/k-mod-bugs.637116/
Also fixed a bug in civics evaluation mentioned there, [advc.001]
and civics are now adopted probabilistically, as had already been the case for religions in AdvCiv. [advc.031]

Made AI less likely to raze the first city conquered on another continent if there is a good chance of conquering further cities on that continent. And some refactoring. [advc.116]

UWAI:
* Revised earlier changes to AI preparations for naval vs. land war.
* Increased impact of the victory-related WarUtilityAspects in the early game. In particular, stage 2 of the Conquest and Domination victory strategies can increase war utility now. (These victory stages had already made the K-Mod AI more warlike, but UWAI had so far only cared about stages 3 and 4.)
* Fixed an issue with AI civs dogpiling on a far stronger opponent when that opponent was already Furious at the dogpile attacker.
* Fixed a bug introduced in v0.93: Military power had been grossly underestimated when a defending civ had cities on several landmasses and a lot of Mounted or other fast units.
* Made the AI a little more optimistic about future uses for the military units that it expects to train for a war. Personality (BuildUnitProb) also matters a bit now. This change makes the AI a bit less worried about the cost of military build-up and a bit more worried about casualties.
* Fixed issues that had prevented the AI from starting a war by invading a vassal.  [advc.104j]

Debug mode: Exiting the Civics screen after switching to the perspective of another civ can no longer change the civics of the human civ. [advc.001]

Merged from the Dawn of Civilization mod:
"Minor civs cannot be involved as the other player in events" [advc.001]
https://github.com/dguenms/Dawn-of-Civilization/commit/d5313deda8960c63f3a1ab755961be64cb2f7e80