Fixed some issues from recent commits + a few AI/UI tweaks DLL included
Fixed an issue with [advc.021b] that had caused the PerfectMongoose map script to assign invalid starting locations after map regeneration.

Fixed an issue with [advc.300] (AI fogbusting): The limit of one guard unit per city site wasn't being enforced.

Amended [advc.144]: The AI considers granting a human request for a gift despite an active peace treaty if the AI remembers receiving a gift from the human player.

Reduced expected embargo duration from 22 to 15 (was 30 in BtS). [advc.130f]

When the Pyramids or Shwedagon Paya is completed, a change-civics popup is shown. [advc.004x]

When a new civic becomes available (through a technology or a wonder) on a turn when no revolution is possible, then the popup is delayed until the player can change civics again. [advc.004x]

And some arbitrary refactoring. [advc.003]