Mostly bugfixes in AdvCiv code (DLL included, still working toward v0.94)
Fixed an inconsistency in [advc.036] (AI resource trades). Might fix this issue:
https://forums.civfanatics.com/threads/advanced-civ.614217/page-5#post-15127487

Fixed an inconsistency in UWAI that had sometimes made the AI unconcerned about falling behind in score.

Corrected an oversight in [advc.179]: The AI evaluation of the production bonus from Apostolic Palace had been skipped when the AI was at stage 0 of the Diplomacy victory.

Fixed a bug introduced by [advc.132] that had caused the button for the force-religion and force-civics espionage missions to be absent sometimes.

The force-religion espionage mission now requires the target religion to be present in the city where the mission is executed. Increased the impact of the number of cities with the target religion on the mission cost. The AI doesn't use the force-religion mission if the target religion has few adherents -- same threshold as the "too few of our people follow that faith" response, and I've tweaked that threshold a bit. (Previously, force-religion missions by the AI had led to a high number of back-and-forth conversions.) [advc.132]

AI no longer incites revolts in undefended cities, nor when the defenders are greatly outnumbered. [advc.120b]

An embargo proposed against a war enemy is forgotten twice as fast (30 turns on average) by that enemy compared with an embargo proposed against a non-hostile target (60 turns on average, as in BtS). [advc.130f]

Loading screen hint added about [advc.124] (naval blockade allows trade to pass through enemy tiles).