More stop-trading tweaks, UI tweaks (combat odds, animations, city alerts), a few minor bugfixes (AdvCiv,K-Mod)
Stop-trading resolutions (UN, AP) cancel recently signed deals, and such deals don't prevent the resolution from being proposed. Exception: The civ that proposes the resolution mustn't have deals with the target that are too recent to be canceled.
AI civs vote against the resolution if they have deals that are too recent. [advc.130f]
(This behavior is consistent with the changes to war resolutions [dlph.25] in a previous commit and is consistent with regular embargoes canceling recently signed deals.)

The AI refuses to stop trading ("not right now") when this would cancel reparations paid for a peace treaty. [advc.130f]

Amended AI refusal to stop trading: No longer fear the "military might" of a vassal. [advc.130f]

Bug (sign error) fixed in a recent change to attitude thresholds for embargoes. [advc.130f]

Fixed a (likely) UI bug in K-Mod code that had sometimes caused a neutral unit to be shown on the map instead of an enemy unit when neutral and enemy units occupied the same tile (making the enemy unit easy to miss). [advc.001]

Fixed a problem with oscillation of tile ownership in [advc.099b].
For performance reasons, the total culture in a tile is now stored in savegames and a function getTotalCulture replaces countTotalCulture in the CvPlot class. [advc.003b]

Improved the AI evaluation of deals where the AI receives two health resources or two happiness resources from a human. (The AI still refuses to trade more than two.) [advc.036]

Changes to combat odds help text: [advc.048]
* Odds help text shows which unit (short description incl. current and maximal strength and promotions) is about to defend and, if multiple units are selected, which one is about to attack.
* Hitpoints of the combatants are no longer shown separately.
* Advanced Combat Odds (ACO) already had options for showing the attacking and defending unit. These are now enabled by default. ACO is still disabled by default.
* Replaced a few uses of the word "retreat" with "withdraw" in combat odds and other help text. ("Withdraw" is the proper term.)

Combat animations become shorter with each era. Configurable through GlobalDefinesAlt.xml. [advc.002m]

Fixed an oversight in AdvCiv code that had prevented the declare-war popup from appearing when a player attempted to attack an enemy unit on the territory of a non-friendly third party. [advc.001]

The "pending" alerts only trigger when a negative event is about to happen: when a city is about to become unhappy or unhealthy or is about to shrink. About-to-become-unhappy now enabled by default (this only affects new installations of AdvCiv). [advc.106d]