XML cache disabled [advc.003i]
It appears that the cache is never written when a mod is loaded, meaning that the cache accomplishes nothing and the write/read attempts only waste time (though probably very little). On top of that, it's possible that, in rare cases, cache data written by the unmodded game is read by the mod, which will result in errors.

Credit to Nightinggale for telling me me how to disable it:
https://forums.civfanatics.com/threads/avoid-creating-bugs-from-the-exe-calling-the-dll.639167/#post-15297196