More options moved from XML to BUG menu; minor changes to resource placement
When checking whether a randomly discovered resource can be placed on a tile, the tile's features are now ignored, meaning that Silver and Gems can be discovered on Grassland Hills. [advc.129,advc.rom3]

Banana, Sugar and Ivory can be placed on Plains Jungle (in addition to Grassland Jungle). This is an addendum to [advc.021a], which allows map scripts to place Jungle on Plains (if Plains are found in appropriately low latitudes).

AI trade value of Marble and Stone increased a bit. [advc.036]

Moved some more options from GlobalDefines_advc.xml to the BUG menu. The options are under a new heading "Layers" on the "Map" tab:
- "Scores in Globe View," "Resource Filters," "Tribal Village Icons" [advc.004z]
- "Founding Yields," "Founding Border" [advc.004h]

Replaced the nonsensical tooltip of the Globe View button. [advc.004g]