Cache of the WarEvaluator class improved
Cleaned up the implementation of the WarEvaluator cache. The cache is now always used when the Trade screen is up, and I've increased its capacity. These changes should eliminate delays that had sometimes occurred in the late game when trading with AI civs willing to declare war on multiple targets. [advc.104l]