Wake up cargo when a ship reaches land
When a ship reaches land with its last movement point, units in cargo are woken up so that the player can't forget to unload them. Units that have been explicitly set to sleep or fortify don't wake up; to implement that, I've added an activity type "boarded" that units adopt upon entering a ship. [advc.075]