Set AI Auto Play handicap to the average of AI handicaps
This commit is a bit of a dud. I wanted to see if I could get the mod to load K-Mod and BtS savegames – no way – and kept a few tiny changes made in the process:

Fixed an old problem with [advc.127] (handicap during AI Auto Play): In scenarios that assign unequal handicaps to the AI civs, the Auto Play handicap is now set to the average of the AI handicaps.

Comment about the SAVE_VERSION in GlobalDefines changed. [advc.009]

Added a parameter for suppressing cache updates when calling setDefineINT. Unused for now. [advc.003b]

Another unsuccessful experiment: I've tried to display resource icons as yield symbols, and found out that the yield symbol paths in XML are actually unused. Added some comments about that.