Alt key for finding attacker with best odds; and a few tweaks.
When a human player attacks or views combat odds with multiple units selected, the unit with the highest survival odds is chosen as the attacker if the Alt key is held down. If Alt is not held, the AI may select an attacker with suboptimal odds, e.g. to inflict collateral damage or to preserve high-level units. [advc.048]
(In BtS, the attacker with the best odds was -usually- displayed, but, when the attack was ordered, the AI was free to choose a different unit.)
Refactored CvSelectionGroup::groupAttack and CvSelectionGroupAI::AI_getBestGroupAttacker a bit in the process.

The change in unit expenses predicted by [advc.004b] now takes into account units in cargo when hovering over the Delete button of a transport unit.

Research overflow is shown as part of the solid portion of the research bar only if tick marks are disabled. With tick marks, it's shown as part of the semi-transparent portion. [advc.078]
The way BtS displays overflow (always solid) suggests that the overflow has already been spent, but it isn't spent until the end of turn. That's also inconsistent with the production bar.

Revolt probability is displayed as 0.1% (instead of 0.0%) even if the actual probability is a bit smaller than 1 permille. [advc.101]

Tweaked a recent change [advc.144] so that the AI may
* ask for a gift despite a peace treaty if 3 or fewer turns of peace remain, and
* demand tribute if 1 turn of peace remains.