Don't show production turns during disorder (and several small tweaks)
Tech cost modifiers for world size adjusted to recent changes (based on watching a few games on AI Auto Play). [advc.910]

The map generator places Horse and Cow only on tiles with a latitude of 9 or greater. On Standard-size Fractal maps, this affects only 4 rows of tiles. [advc.129]
(To account for the tsetse fly. Pig and Sheep already had a latitude restriction.)

Tweaked the probability for decreasing the occupation timer so that the timer can decrease without an occupying force when not at war with the revolt player. [advc.023]
(Don't want a revolt to continue forever when the city owner refuses to move any units into the city.)

During disorder, the city UI doesn't show the number of remaining production turns. (BtS treats cities in disorder as having a production rate of 1 when computing the number of remaining turns.) I had already changed this for research turns in an earlier commit. [advc.004x]

When starting in a later era, a change-civics popup is shown on the start turn. [advc.004x]

Changed help text for the "Load" button when the selected unit is already in cargo. Now says that the unit will be transferred to another ship. [advc.004g]

When playing with the "Always War" option, UWAI no longer shows enemies as willing to talk on the scoreboard. [advc.104i]

Removed the single-player-only hint from the help text of the "Lock Modified Assets" and "Rise & Fall" game options because these options no longer appear on the Staging Room screen. [advc.054,advc.701]

A few cosmetic changes in OOSLogger.py.