Minor changes to Advanced Start (from Kek-Mod)
Reviewed the recent Kek-Mod changes to Advanced Start and adopted the following:
* Make sure Barbarians don't receive any start points or free units.
* Start point penalty for the AI on difficulty levels below Noble.
https://github.com/DarkLunaPhantom/Civ4-K-Mod/commit/b7e8b045aef19374b0148203bcf758d54c296549
https://github.com/DarkLunaPhantom/Civ4-K-Mod/commit/04ed5df43da1449208cfd673b2b04a5bc2b58dff

Plus some minor style changes to [advc.250c] and the restriction (added by me) that Ocean tiles can't be revealed in Advanced Start unless adjacent to Coast no longer applies in the Industrial era and later.

And an adjustment of tech costs on Noble difficulty that I had meant to commit earlier this month but forgot (apparently). [advc.251]