Costs for Space victory reduced
Reduced tech costs in the Modern and Future era and reduced cost modifiers for team projects on Emperor difficulty and above. And improved the game text for the global PROJECT_PRODUCTION_PERCENT modifier, but decided to leave it at 100% in the end. [advc.251]
(Previous cost increases, especially those based on the difficulty level, had made Space victories even more grueling than in BtS.)