Full-screen Sevopedia (work in progress)
Tbd.: Use the additional vertical space to enlarge the panels that need it most. Currently, only the bottommost panel on each page is enlarged.

And, under "Shortcuts", moved Alt+Z from the "Mods" subcategory to "Game" because it's a BtS shortcut. And added Alt+Shift+Z.

(change id [advc.004y])