Exported functions with inline counterpart renamed via .def file
change id: advc.003f (but manual still to be updated)

Inspired by posts by alberts2 and Ang in the Caveman2Cosmos forum --
https://forums.civfanatics.com/threads/modders-documentation.441325/page-16#post-15476563
-- I've added a module definition file that renames functions that the DLL has to export but shouldn't call internally: getX, getY, getOwner (in various classes), getGame, getMap and several CvMap functions. The new function names all have the suffix "External". The DLL should use the inlined versions instead.

I've appended "INLINE" to numerous call locations - however, the next step (and the real point of this commit) is to remove the "INLINE" suffix everywhere to unclutter the code. This will be easy to do (automatic search and replace), but I want to wait a few days since this .def file thing is still new to me.

I've already removed the "INLINE" from CvSelectionGroup::getOwnerINLINE and CvPlotGroup::getOwnerINLINE because the external versions of those functions weren't even called by the EXE.

As for exported functions that miss a const qualifier, I've realized that a const function with the same name works just fine; no need to rename the original function through the .def file. So I've renamed my const function CvGame::gameTurn to CvGame::getGameTurn and used the same pattern for a few other (unimportant) functions. [advc.003]

The .def file is enabled through Makefile.project. And I've also moved the /WX compiler flag to Makefile.project. [advc.make]

(And, unrelated: commented out a bit of unused code in CvMapGenerator.py)