Renamed getX_INLINE, getY_INLINE to getX, getY
Through automated search and replace, replacing "_INLINE" with an empty string. Then manually changed some linebreaks (now that the function names are shorter, getX and getY should always be on the same line when both are passed as function arguments) and removed some stray whitespaces that I happened to come across.

Follow-up on this commit:
https://github.com/f1rpo/AdvCiv/commit/6c424b0e300fc3b5c498ad0c1177c1b58c0dde5d

Coming up: Rename the remaining "INLINE" functions.