Break ties in resource placement order randomly
[change id advc.129]

During map generation, when the same placement priority is set for two types of resources in XML, decide randomly which type to place on the map first.

Similarly, during starting plot normalization, go through the resource types in a random order when looking for a suitable food or extra resource.

BtS relies on the order of the XML file (mostly alphabetic) in these cases. That reduces the variety of starting locations and causes Sugar (one of the last resources in the file) never to be placed on extremely crowded maps. With my changes, some resource type may still be omitted, but it won't always be the same type.

These changes were prompted by this thread on CFC:
https://forums.civfanatics.com/threads/always-missing-specific-resources-for-small-maps-islands-custom-continents.648468/