Split up CvInfos.h
[change id advc.003x]

To make the code for accessing XML data easier to work with, to reduce compilation time and to encourage encapsulation.
For rationales for this particular way of partitioning CvInfos.h, see the comments in the new CvInfo_*.h files.

I've ended up with a total of 18 headers. Some of these include each other though and many are included in CvGameCoreDLL.h; so the the overall number of include directives in the code hasn't increased much.

I've removed the unused iInfoBaseSize parameter from the CvXMLLoadUtility::SetVariableListTagPair functions to remove some unnecessary dependencies on the info classes.

Within the info classes, I've taken this opportunity to declutter the code a bit:
* Removed unhelpful comments (i.e. most comments).
* Removed empty default constructors and destructors.
* Replaced pairs of FAssertMsg calls with a single FASSERT_BOUNDS call.
* Replaced if/else sequences with switch statements in the implementation of CvActionInfo.
And I've inlined some functions. [advc.003f]

Minor (mostly) unrelated changes made along the way:
* Addressed problems with CvGlobals::getInfoTypeForString receiving arguments (from I don't know where) that refer to enum types rather than info types by having getInfoTypeForString fall back on CvGlobals::getTypesEnum. [advc.006]
* When a unit is inside a rival city, the 3D models of all city buildings become visible to all members of the unit's team. Previously, they were only visible to the unit owner. [advc.045]
* Added some more bogus declarations to CvGameCoreDLL.h to avoid boost/python errors in the VS code editor. [advc.make]
* A few random style changes (not many). [advc.003]