Moved DLL-to-Python calls and logging into separate classes
New class CvPythonCaller. [advc.003y]
To make (unused) Python callbacks less visible in the DLL code. Also improves error checking: by default, CvPythonCaller checks whether the Python function was successfully called. Found out about two (unused) missing Python functions this way; removed the one (doHeadquarters) on the DLL side, added the other (cannotSpreadReligion) on the Python side. This has required me to copy CvGameInterface.py from BtS.

CvPythonCaller also handles the caching of the callback guards loaded from PythonCallbackDefines.xml. This is now done through an array of enum values. I've added some more callback guards, but the gains in performance are negligible (1% or so). The guards for Python events are now handled by the CvDllPythonEvents class.

Moved the computation of game score and capture gold from Python to the DLL (CvPlayer). Not sure why these had been implemented in Python; perhaps as instructional examples for Python modders. I might want to change those formulas though, and that'll be easier (for me) in C++ (and faster).

Similarly, I've moved the code that triggers the "Partisans" event from Python to the DLL (CvCity) and slightly changed it so that units spawn for the civ with the highest tile culture instead of the one with the highest city culture.

New class CvDLLLogger. [advc.003t]
To encapsulate the "message (control) log". That's used mainly for logging combat and random numbers (if "RandLog" is enabled in addition to "MessageLog").

Changes made along the way – mostly style/ minor refactoring. [advc.003]
* Disabled some more log output in CvEventManager.py by default. [advc.007]
* The minimal distance between starting locations is now computed only once (by CvGame::updateStartingPlotRange) instead of computing it for every possible starting location in CvPlayerAI::AI_foundValue. [advc.003b]
* Removed the empty public CvRandom::uninit function.
* Removed some unhelpful Firaxis comments from init and reset functions.
* Removed "public functions ..." and "protected functions ..." comments from most implementation files (not helpful in such large files).
* Utility function "intToShort" added.
* Utility function "narrowUnsafe" added for converting wide character strings to narrow character strings. Doesn't take into account locale and doesn't warn about characters that can't be narrowed – but at least these potential problems are now communicated by the function name. It's only used for map script names (so far).
* Some refactoring in CvTeam::setHasTech and (related) in CvGame::doHolyCity and CvGame::doHeadquarters. Two bugfixes in unused code in doHeadquarters.
* Refactoring in CvGame::testVictory.
* Fixed a linker problem with psapi.h more properly (I hope). [advc.make]