Changed names of loop counters in some K-Mod code
This old K-Mod commit
https://github.com/f1rpo/AdvCiv/commit/c463b0d89baac8e4e4ae45afb44e3e97220a2f7f
convinced me that loop counters really shouldn't be given names like "ePlayer" that are commonly used for other kinds of variables. The problem is that MSVC03 gives loop variables a scope beyond the body of the loop; see this CFC post of mine:
https://forums.civfanatics.com/threads/cvgamecoredll-build-error-boost-python-list-hpp.648072/#post-15509430

So I've changed all K-Mod loops where a variable "iPlayer", "ePlayer", "iTeam" or "eTeam" was declared in the loop header. Loops of the same pattern still exist – iterating over the various enum types –, but it's not worth the trouble to locate all those.