PerfectMongoose: tweak Plains, Forest frequency; tropics.
Decreased the overall frequency of Plains a little.
Some Plains are now placed on plots with an especially wet climate (i.e. near the equator).
Increased the frequency of Forests a bit.
Snow hills and Tundra can't appear in the tropics; flat Snow can only appear in the polar circles.
Increased land-to-sea ratio a little bit.

Prompted by this discussion with CFC user LeBashar:
https://forums.civfanatics.com/threads/advanced-civ.614217/page-15#post-15535048