PerfectMongoose performance tweak
When placing lakes, the map areas are now recomputed only once in the end. They were previously recomputed after every added water tile, which had added a couple of seconds to the total map generation time. I'm not sure if this is purely a nonfunctional change, but at least I'm not seeing any adverse effects. [advc.opt]

And this change mentioned in the previous commit message ...
"Attempt to turn lakes fed by a river into bays only with a 50% probability." [advc.021b]
... is actually in this commit.