Show custom map options on Settings tab
The values of custom map options are listed on the Settings tab (Victory screen) if they differ from the default. For example, normally, "Cylindrical" (world wrap) is not listed because that's the default value, but "Toroidal" gets listed. [advc.004]

(I've wanted to implement this for some time, but only recently learned how to easily obtain option value descriptions from a map script.)