Fix expanding scoreboard in simultaneous turns
The scoreboard had sometimes stayed permanently expanded. The underlying problem was my code for detecting when the game is in between turns [advc.003r]. Now assumes that the game is never in between turns when turns are taken simultaneously. Hopefully, this doesn't mess up BUG alerts [advc.210]; looks OK so far.