Fix v0.96 bug in AI closeness cache
The cache for the (geographical) closeness between the cities of an AI civ and those of any given rival was not getting updated reliably. In fact, many of the closeness values seem to have remained at 0 (meaning not close at all) for dozens of turns if not forever – a serious impediment for AI war/ strategic planning. I'll include the bugfix in v0.96d.