Partial update of manual
The update only covers the (few) changes in v0.96d. I've already pushed that version of the manual (odt and pdf) to the hotfixes branch, but various links point to manual.pdf on the master branch; so I'm copying the pdf from hotfixes to master. The odt on master is work in progress; covers a few changes that aren't included in v0.96d. (Though most of the changes since v0.96 are only documented through commit messages so far; I'll basically have to copy those to the proper places in the manual.)