Announce city trades to third parties
Show an announcement to third parties when a city is ceded and record the ownership change in the replay log. (Previously, only liberated cities were announced.)

When a ceded city establishes contact between two civs, have them meet before announcing the cession and place the civilization-met indicator at the city tile.

Refactoring in CvPlayer::acquireCity and CvCity::liberate

While I was at it: replaced all occurrences of "(*it)." with "it->"