City trade alert added.
Triggers when a city can be traded, i.e. when another player becomes willing to cede a city or to trade for a city, or when a city can be liberated. Does not trigger for cities that can already be traded when founded or acquired. And a few other exceptions to avoid unnecessary notifications. Hardly tested so far. [advc.ctr]

Simplified the AdvCiv4lerts base class a bit by getting rid of the init and reset functions. And alert objects are only created for the active player now (saves a tiny amount of memory). [advc.210]

Some whitespace changes in CvPlayer::setTurnActive.