Another tweak to the demand-better-protection option
Halve the defensive strength of a unit also if the unit can no longer be trained in the capital. So that players can't increase a city's garrison strength by deliberately removing access to a strategic resource. (Would have to remove the resource from the capital in addition, which isn't normally feasible.)