Fix Ctrl+H (select hurt units)
[change id advc.001z]

When a city or fort tile contains units of different domains (land/ sea/ air), Ctrl+H selects all wounded units with the same domain as the currently selected unit(s). If no units of that domain are wounded, Ctrl+H has no effect.

In Warlords (which introduced the shortcut), the domain was selected based on the internal order of the plot list and if no unit was hurt, all units became unselected and automatic unit cycling (if enabled) proceeded to the next group.

Thanks to CFC user xyx for making me aware:
https://forums.civfanatics.com/threads/advanced-civ.614217/page-19#post-15579814

This commit includes some style changes in CvGame::doControl.