Single popup for canceled deals; rename some parameters.
When an AI civ cancels several resource deals with a human civ at once, only one renegotiation popup is shown. That popup lists the trade items of all canceled deals. [advc.133]
This should also solve a minor issue with showing "you-must-be-joking" resources in the renegotiation popup [advc.074].

Renamed the PlayerTypes/TeamTypes parameters of the AI trade denial and evaluation functions. Previously, those parameters were all named "ePlayer" or "eTeam", however, for the denial functions, the recipient is the parameter, whereas the evaluation functions are called on the recipient.