Fix spurious "can no longer work on" popups
When updating all plot groups during the active player's turn, e.g. after discovering Sailing through a tech trade, it was possible that cities producing an order with a resource requirement were interrupted.

The problem was introduced with [advc.064d] (blocks an exploit for storing arbitrary amounts of overflow production), so I think it can occur in the v0.96 releases; hopefully only rarely.

The new code should also be a little faster because it no longer checks for maxed out wonders when plot groups are updated.

(And a few misc. style changes.)