Don't require RandLog for OOSLog
Only MessageLog=1 is required now. When debugging OOS errors, RandLog=1 should normally be set, but when playing for the sake of playing, RandLog writes too much data. The MessageLog requirement is somewhat important though because the OOS log will contain hidden parts of the game state and MessageLog=1 makes sure that players are aware of that by showing a "logging enabled" popup on the Staging Room screen.

[change id dlph.27]