Use AgentIterators in CvTeam.cpp
CvTeam and CvTeamAI (to be done) contain a lot of loops over all members of a team, and those can be written much more concisely and a somewhat more efficiently with the recently added MemberIter class.

Also added moved some more function definitions to CvTeam.h in order to allow inline expansion.