Two minor bugfixes in AdvCiv AI code
Both bugs are present in the latest releases.

When considering to found a city with a resource on another landmass in its radius, the found-city value was sometimes decreased more than I had intended. [advc.040]

Two (pre-AdvCiv) special conditions that are so supposed to stop the AI from razing cities had been ignored. [advc.116]