Move AI_foundValue computation into a new class
Also split the code into subroutines. This has mandated -or strongly suggested- other refactoring changes to the found value code.

I've taken this opportunity to add logging code. When the log is enabled through BBAILog.h, a breakdown of the found value computation is written to the BBAI log file whenever an AI player founds a city and when placing or normalizing a starting location. When an AI city is founded, breakdowns are also written for the found value of the best adjacent tile and for the next best city site overall. [advc.031c]

The few logs I've looked at and the refactoring changes have brought to light or reminded me of some minor issues, which I've fixed: [advc.031]
* Improved evaluation of chopping (was previously factored into the health evaluation).
* Revised tech requirement checks for resources (techs for revealing, improving, trading the resource).
* Removed a K-Mod kludge (negative yields assumed on the city tile) that was supposed to discourage the AI from settling on top of resources and flood plains. Hopefully no longer needed.
* Disabled K-Mod code that was supposed to steer the distribution of starting locations among landmasses. As far as I can tell, CvPlayer::findStartingArea handles that.
* Made the AI more pessimistic about flipping tiles near a colony on a continent where a rival already has multiple cities.
(Thanks to keldath for making me aware of that potential problem a while ago.)
* Don't count on getting the upper hand culturally when both sides have the Creative trait.
* Disabled some minor (probably inadvertent) visibility cheats.
* Made the AI reluctant to move its starting Settler more than 1 tile on maps with a higher-than-normal player count (e.g. more than 8 on a Standard-size map with Medium sea level). To make situations like the one reported here by Elkad
https://forums.civfanatics.com/threads/advanced-civ.614217/page-22#post-15602780
more unlikely. [advc.108]
* Slightly increased production yield weight for Barbarian cities (no bias previously). [advc.303]
* Some tweaks in the evaluation of sea access.
* Minor bugfix: AI personality had sometimes affected city site evaluation when computing recommendations for humans and when placing starting locations.

Only tangentially related:
Choosing the "Balanced resources" map option now causes more aggressive normalization of starting locations. In particular, unrevealed resources are then taken into account. [advc.108]