Allow air unit production in cities with insufficient capacity
[change id advc.001b]

Basically back to how it worked in BtS. The AdvCiv rule can still be enabled through GlobalDefines_advc.xml.

Unlike in BtS, units that are force-moved upon completion have their movement points spent. When a (legal) rally point is set, force-moved units are sent there. When no rally point is set, units are moved to the nearest legal city or fort (or destroyed if there is none) as in BtS and a message is shown on the main interface.