Change culture garrison strength formula; add tooltip.
Culture garrison strength is now equal to current combat strength with the following modifiers:
- Defensive modifiers (Archery units, Garrison promotions)
- Generic strength modifiers (Combat promotions)
- Minus 50% for non-lethal units (i.e. Siege except Machine Gun).
iCultureGarrison in XML is obsolete.

Adjusted the foreign culture strength formula to balance out the change to culture garrison strength, and to make revolts a bit easier to suppress in the midgame and a bit harder in the late game. Added two XML parameters for fine-tuning this.

Added hover text on the main map and city screen (nationality bar) showing how much more garrison strength is needed to reduce a city's revolt chance to 0, or how much garrison strength can be safely removed.