AI may pay for peace even if it thinks that both sides want it
(UWAI component)

Previously, an AI civ had never paid a human player for peace if it thought that peace was in both sides' interest. Now the AI will pay based on the difference between AI peace utility and two times the human peace utility. (It won't pay if that difference isn't positive.)

Thanks to everyone on the CFC forums who had complained about the old behavior. It was too reliant on what the AI "thought" about the human player's situation, and it's not a reasonable position to insist on a square peace treaty when continuing the war is disastrous for one side and merely inconvenient for the other.

The AI is now also a bit more generous when paying for peace in situations when the war is going badly but not extremely badly. (For extremely bad wars, the AI still assumes that peace is unlikely to last and therefore not worth very much.)