Fix BBAI bugs in automated air recon
In the code that selects the recon target tile, a visibility check was flipped. As a result, the target was chosen randomly with a bias only for maximal distance from the airbase.

Also, tiles around the airbase had been evaluated instead of tiles around the candidate recon tile.

BBAI uses its AI air recon function only for automated human units, but change [advc.650] also lets the AI use it.

Even with the bugs fixed, it's not a good heuristic. I'll make some tweaks in a separate commit.