Remove global getAsyncRand function
This reverts part of SHA 3d9626f47961e52db72e27138533ecc8b3f1f6d0

Since CvGlobals has only a pointer to the non-synchronized PRNG, there is no problem with const-correctness – and so that (truly) global accessor function was an unnecessary addition.

(I'm in the process of filling some gaps in the documentation part of the manual.)

Also: My general cleanup of CvGlobals is no longer tagged with "advc.003t", but just with "advc" or "advc.opt".