Class scaled_int for fixed-point math incl. pow function
[change id advc.fract]

Some initial tests for correctness and speed look promising. Seems like scaled_int won't be significantly slower or faster than double (nor did I expect it to be faster). Not sure that I've overloaded all the right operators. Perhaps I'll switch to using boost/operators.hpp later.

The power function (for fractions) could surely be improved in terms of accuracy and speed, but, for my needs, the current implementation may well be sufficient.

I don't think I'll convert all my floating-point code to scaled_int, or at least not the UWAI component. The interface of scaled_int is intended to be rich enough though to make that conversion feasible if necessary.

I intend to use scaled_int in modified BtS code in order to make my changes portable to mods that don't use floating-point numbers. ScaledInt.h itself should be pretty easily portable.
(But I won't change the rounding behavior in places that players can and do observe, e.g. +25% from Forge will continue to amount to 0 when the city's base production rate is less than 4.)

Ties in with this "We the People" issue (and #169):
https://github.com/We-the-People-civ4col-mod/Mod/issues/176