Treat NoWarProb of Random Pers. as secret; map trade speed tweak.
The AI now treats the NoWarProb values of other AI leaders as secret when playing with Random Personalities. [advc.104y]

When trying to even out an AI-AI tech trade, maps are only considered when they have a significant trade value. [advc.136b]
This was intended as a performance tweak, but it turns out that by far the slowest part of a map trade is the graphics update – and that occurs only when the active human player is involved. Specifically, the update of route graphics is very slow; it seems that the game core DLL can't do anything about that.

And some misc. refactoring.