Tweak impact of city population on revolt chance
Apply the era-based cap on the population factor _after_ shifting it toward the highest-ever population. And shift the population factor back toward the actual population eventually so that cities that stay small in the long run don't indefinitely get a revolt chance based on the highest-ever population.

Prompted by this post by CFC user Lanstro:
https://forums.civfanatics.com/threads/advanced-civ.614217/page-29#post-15723888