Tweak revolt chance penalties, nationality bar help
[change id advc.101]

Indicate hurry anger, draft anger, foreign religion (if not being countered by owner's state religion) in the hover text of the nationality bar. The text doesn't say by how much these factors increase foreign culture strength; too complicated to implement.

Decrease the impact of hurry and draft anger. Previously, a city was treated as having 5 extra population while angry. Now it's 1.5 extra population per 1 unhappiness (from hurrying or drafting). And the 50% overall increase in foreign culture strength when angry is now adjusted to game speed (because the anger duration is longer in slower games) and taken times the square root of the unhappiness.

Reduced the impact of religion in cases when neither the city owner nor the cultural owner have a state religion or when neither of those religions is present in the city.

Instead of player tech era or game era, revolt chance now scales with a value computed from the percentage of techs discovered by cultural owner and city owner. That value gets quantized in order to avoid frequent changes in revolt probability.

Related discussion on CFC:
https://forums.civfanatics.com/threads/advanced-civ.614217/page-31#post-15735516