Bugfix in AI proposals for dual deals
When asked to make a deal work that has 0 trade value for both sides, the AI should say that the deal is fine as it is. It had been saying that no deal is possible. In particular, the AI sometimes wouldn't propose a square peace treaty.

The bug was introduced in some recent commit.

Thanks to SuperXANA reporting the issue and providing a savegame.