Fix AdvCiv bug: Barbarian Galley stuck in a loop
When Barbarians had appeared aboard a Galley (through [advc.306]), and that Galley was unable to reach any city because of ice and tried to drop its cargo at the nearest land tile, the Galley got stuck in a loop if the nearest land tile happened to be a peak. Looks like this only caused the Galley to miss one turn. It did trigger an assertion though.