Allow "enough-on-our-hands" indicator when UWAI disabled
[change id advc.104]

The war trades alert doesn't really render the scoreboard indicator obsolete. Players can enable it by adding the letter 'M' to the scoreboard string on the BUG menu. This won't have any effect when UWAI is enabled.

I've modified the logic for showing the indicator so it only appears next to civs that aren't already at war.