Let friendly Fort w/o route connect water tiles
[change id advc.124]

Tweak (bugfix?) for the trade network rules. As shown in this screenshot by CFC user crullerdonut
https://forums.civfanatics.com/attachments/no-trade-routes-allowed-jpg.562884/
a Fort without a route did not connect water tiles. The screenshot is from AdvCiv, but I think it works the same way in BtS.

That's counterintuitive; ships don't travel on roads and railroads. A route is still required in order to connect a land resource on the Fort tile.