Don't place unreachable water resources
Water tiles that could only be worked from a city on an impassable tile (Peak) are counted as unworkable, meaning that they have 0 yield and no resources can be placed there. [advc.129d]

Not covered: Water tiles workable by a city - but fully enclosed by Ice and thus not reachable for Work Boats. Such tiles shouldn't receive resources, but that's not so easy to prevent.